requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.3.2
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
parse==1.20.0
pluggy==1.3.0
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
requests-wsgi-adapter==0.4.1
urllib3==2.1.0
waitress==2.1.2
WebOb==1.8.7
whitenoise==6.6.0
